+++ agents-flex-search-engine/agents-flex-search-engine-lucene/src/main/java/com/agentsflex/search/engine/lucene/LuceneSearcher.java
@@ -164,7 +164,7 @@ public class LuceneSearcher implements DocumentSearcher {
                 resultDoc.setTitle(doc.get("title"));
                 resultDoc.setMetadataMap(extractMetadata(doc));
 
-                resultDoc.setScore((double) scoreDoc.score);
+                resultDoc.setScore(scoreDoc.score);
 
                 results.add(resultDoc);
             }
